make things pretty
just gonna review my own code yanno
