+++ templates/game.html
@@ -1,9 +1,9 @@
-<!--user.html-->
+<!--game.html-->
 
 <!doctype html>
 <html>
 
-  <head>
+<head>
     <meta charset="UTF-8">
     <title>Game with {{ range .playerNames }} {{ . }} {{ end }}</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -13,36 +13,44 @@
             height: 5rem;
             width: 4rem;
             text-align: center;
-            /* vertical-align: center; */
             border: 2px solid black;
         }
+
         .card.unknown {
             /* TODO make this be the cool bicycle photo */
-            background: black;  
+            background: black;
         }
+
         .card.cut {
             background: gray;
         }
+
         .card.pegged {
             opacity: 0.6;
         }
+
         .card.opponent:not(.unknown) {
             /* known opponent cards should be gray */
             background: gray;
         }
+
         .card.disabled {
-            background: lightgray;  
+            background: lightgray;
         }
+
         .card.activated {
             border: 2px solid red;
         }
+
         #container {
             margin: auto;
         }
+
         #summary {
             width: fit-content;
             float: left;
         }
+
         #interaction {
             width: fit-content;
             float: left;
@@ -51,77 +59,85 @@
         }
     </style>
 
-  </head>
+</head>
 
-  <body>
-      <div id="container">
-      <div id="summary">
-    <h1>Game</h1>
-    <h4>Players: <br>{{ range .playerNames }} - {{ . }}<br> {{ end }}</h4>
-    <div>Current dealer: {{ .currentDealer }} </div>
-    <div>Current phase: {{ .phase }} </div>
-    <h2>Scores</h2>
-    <div>
-    <div>You are: {{ .myColor }} </div><br>
-    {{ range .scores }}
-        <a> {{ .Color }} :  {{ .Score }}</a><br>
-        {{ end }}
-    </div>
+<body>
+    <div id="container">
+        <div id="summary">
+            <h1>Game</h1>
+            <h4>Players: <br>{{ range .playerNames }} - {{ . }}<br> {{ end }}</h4>
+            <div>Current dealer: {{ .currentDealer }} </div>
+            <div>Current phase: {{ .phase }} </div>
+            <h2>Scores</h2>
+            <div>
+                <div>You are: {{ .myColor }} </div><br>
+                {{ range .scores }}
+                <a> {{ .Color }} : {{ .Score }}</a><br>
+                {{ end }}
+            </div>
 
-    {{ if .cutCard }}
-    <h2>Cut card</h2>
-    <div>
-        <div class="card cut">{{ .cutCard }}</div>
-    </div>
-    {{ end }}
+            {{ if .cutCard }}
+            <h2>Cut card</h2>
+            <div>
+                <div class="card cut">{{ .cutCard }}</div>
+            </div>
+            {{ end }}
 
-    {{ if .peggedCards }}
-    <h2>Pegged Cards</h2>
-    <h3>Current Peg: {{ .currentPeg }}</h3>
-    <div>
-        {{ range .peggedCards }}
-        <div class="card pegged">{{ . }} </div>
-        {{ end }}
-    </div>
-    {{ end }}
-    
-    {{ if .myHand }}
-    <h2>Your hand</h2>
-    <div>
-        {{ range .myHand }}
-        <div class="card mine {{ if .IsPegged }}disabled{{ end }}" id="{{ .Card }}" >{{ .Card }} </div>
-        {{ end }}
-    </div>
-    {{ end }}
+            {{ if .peggedCards }}
+            <h2>Pegged Cards</h2>
+            <h3>Current Peg: {{ .currentPeg }}</h3>
+            <div>
+                {{ range .peggedCards }}
+                <div class="card pegged">
+                    {{ . }}
+                </div>
+                {{ end }}
+            </div>
+            {{ end }}
 
-    {{ if .oppHands }}{{ range .oppHands }}
-    <h2>{{ .Name }}'s hand</h2>
-    <div>
-        {{ range .Hand }}
-        <div class="card {{ if not .IsKnown}}unknown{{ end }} opponent" >{{ .Card }} </div>
-        {{ end }}
-    </div>
-    {{ end }}{{ end }}
+            {{ if .myHand }}
+            <h2>Your hand</h2>
+            <div>
+                {{ range .myHand }}
+                <div class="card mine {{ if .IsPegged }}disabled{{ end }}" id="{{ .Card }}">
+                    {{ .Card }}
+                </div>
+                {{ end }}
+            </div>
+            {{ end }}
 
-    {{ if .crib }}
-    <h2>Crib</h2>
-    <div>
-        {{ range .crib}}
-        <div class="card {{ if not .IsKnown}}unknown{{ else if not $.myCrib }}opponent{{ end }}" >{{ .Card }} </div>
-        {{ end }}
+            {{ if .oppHands }}{{ range .oppHands }}
+            <h2>{{ .Name }}'s hand</h2>
+            <div>
+                {{ range .Hand }}
+                <div class="card {{ if not .IsKnown}}unknown{{ end }} opponent">
+                    {{ .Card }}
+                </div>
+                {{ end }}
+            </div>
+            {{ end }}{{ end }}
+
+            {{ if .crib }}
+            <h2>Crib</h2>
+            <div>
+                {{ range .crib}}
+                <div class="card {{ if not .IsKnown}}unknown{{ else if not $.myCrib }}opponent{{ end }}">
+                    {{ .Card }}
+                </div>
+                {{ end }}
+            </div>
+            {{ end }}
+        </div>
+        <div id="interaction">
+            <button id="refreshGame">Refresh</button><br>
+            <button disabled id="dealButton">Deal</button><br>
+            <button disabled id="buildCribButton">Build crib</button><br>
+            Cut Percent: <input disabled type="number" name="cut percent" id="cutInput"><br>
+            <button disabled id="pegButton">Peg</button><br>
+            Hand Points: <input disabled type="number" name="hand points" id="handPtsInput"><br>
+            Crib Points: <input disabled type="number" name="crib points" id="cribPtsInput"><br>
+        </div>
     </div>
-    {{ end }}
-</div>
-    <div id="interaction">
-        <button id="refreshGame">Refresh</button><br>
-        <button disabled id="dealButton">Deal</button><br>
-        <button disabled id="buildCribButton">Build crib</button><br>
-        Cut Percent: <input disabled type="number" name="cut percent" id="cutInput"><br>
-        <button disabled id="pegButton">Peg</button><br>
-        Hand Points: <input disabled type="number" name="hand points" id="handPtsInput"><br>
-        Crib Points: <input disabled type="number" name="crib points" id="cribPtsInput"><br>
-    </div>
-</div>
 
     <script src="/assets/js/wasm_exec.js"></script>
     <script>
@@ -131,19 +147,20 @@
         let mod, inst, bytes;
         console.log("initializing gowasm");
         WebAssembly.instantiateStreaming(
-          fetch("/assets/wasm/wa_output.wasm", 
-            // change default to 'no-cache' while deving to force getting it
-            {cache: 'default'}), 
+            fetch("/assets/wasm/wa_output.wasm",
+                // change default to 'no-cache' while deving to force getting it
+                { cache: 'default' }),
             go.importObject).then((result) => {
-          mod = result.module;
-          inst = result.instance;
-          memoryBytes = new Uint8Array(inst.exports.mem.buffer)
-          console.log("initialized gowasm");
-          run();
-        });
+                mod = result.module;
+                inst = result.instance;
+                memoryBytes = new Uint8Array(inst.exports.mem.buffer)
+                console.log("initialized gowasm");
+                run();
+            });
         async function run() {
-          await go.run(inst);
+            await go.run(inst);
         }
-      </script>
+    </script>
 </body>
+
 </html>
