+++ wasm/consts/elementIDs.go
@@ -1,5 +1,13 @@
 package consts
 
+// home page
+const (
+	CreateUserButtonID       string = `createUserButton`
+	CreateUsernameInputID    string = `createUN`
+	CreateDisplaynameInputID string = `createUN`
+)
+
+// game page
 const (
 	RefreshButtonID string = `refreshGame`
